requirements.txt
click==8.0.4
colorama==0.4.4
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==2.11.3
MarkupSafe==2.0.1
Werkzeug==2.0.3
pillow==9.0.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.0
pytorch-lightning==2.0.0
torchvision==0.15.1